feat: restructure project modules
