+++ src/main.ts
@@ -1,7 +1,7 @@
 import { ValidationPipe } from '@nestjs/common';
 import { NestFactory } from '@nestjs/core';
 import { ConfigService } from '@nestjs/config';
-import { AppModule } from './app.module';
+import { AppModule } from './modules/app/app.module';
 import { setupSwagger } from './swagger';
 
 import pinoHttp from 'pino-http';
